step4: refactor: 부울, 부울목록, 사다리목록 생성기 이름 변경
