+++ src/main/java/generator/bools/impl/NonContinuousGenerator.java
