--- src/main/java/generator/ladder/LaddersGenerator.java
@@ -1,7 +1,0 @@
-package generator.ladder;
-
-import domain.ladder.Ladders;
-
-public interface LaddersGenerator {
-    Ladders generate(int height, int size);
-}
