+++ src/main/java/generator/ladders/LaddersGenerator.java
@@ -1,0 +1,7 @@
+package generator.ladders;
+
+import domain.ladder.Ladders;
+
+public interface LaddersGenerator {
+    Ladders generate(int height, int lineSize);
+}
